refactor: Update NavbarActions and Search components
This commit refactors the NavbarActions and Search components to improve the user experience and add new functionality. In the NavbarActions component, the code has been updated to handle the case when the component is not mounted. In the Search component, the code has been updated to include a search icon and improve the styling of the search input. These changes enhance the usability and visual appeal of the components.