refactor: Add colorNameVisible field to Product model
The `colorNameVisible` field has been added to the `Product` model in the `prisma/schema.prisma` file. This field is set to `false` by default and allows for the visibility of the color name for a product.

This change enhances the flexibility of the `Product` model by providing the ability to control the visibility of the color name.